Moved to day1 package
