+++ src/main/kotlin/aoc2015/solution/day1/001NotQuiteLisp.kt
@@ -1,4 +1,4 @@
-package aoc2015.solution
+package aoc2015.solution.day1
