+++ src/main/kotlin/aoc2015/solution/day1/001NotQuiteLispPart2.kt
@@ -1,4 +1,4 @@
-package aoc2015.solution
+package aoc2015.solution.day1
 
 
 fun notQuiteLispPart2(instructions: String): Int {
